Implement MQTT Alert Handler
Much of this code was pulled from the guide published here:
https://github.com/influxdata/kapacitor/blob/master/alert/HANDLERS.md .

This also adds eclipse/paho.mqtt.golang as a dependency to Kapacitor.
This client was chosen primarily because it is the same one that is used
in Telegraf. It is actively developed, and is widely used by other
popular projects in the community (e.g. Gobot).