Set CleanSession to true for MQTT
Clean session, as described in the comment, is a parameter which
controls whether or not the broker holds onto the session state for a
connecting client. This is useful in scenarios where clients are
resource-constrained. For example, it could be quite expensive to push
long topic strings onto a broker at boot-time.

Since Kapacitor is typically not resource-constrained, and is only
publishing messages, it's safe to hardcode this to true. This conclusion
was derived from the best practices section here:
http://www.hivemq.com/blog/mqtt-essentials-part-7-persistent-session-queuing-messages